Added hacked up arrow code

+++ core/src/com/sillygames/killingSpree/networking/messages/EntityState.java
@@ -2,7 +2,7 @@ package com.sillygames.killingSpree.networking.messages;
 
 public class EntityState {
     
-    public byte id;
+    public short id;
     public byte type;
     public float x, y;
     public short angle;
